Avoid ambiguous implicit errors with dotty
For a Boolean b and a prop p, the following is ambiguous in dotty:
b ==> p
Because we have:
  implicit def BooleanOperators(b: => Boolean) = new ExtendedBoolean(b)
  implicit def propBoolean(b: Boolean): Prop = Prop(b)

And both ExtendedBoolean and Prop provide a `==>` method, the one in
`ExtendedBoolean` just forwards to the one in `Prop`. This is not
ambiguous in scalac because `Boolean` wins against `=> Boolean` but it
is in dotty.

This commit fixes this by removing the `==>` method from
`ExtendedBoolean`, this may break code that explicitly imported
`BooleanOperators` but not `propBoolean`, the fix is to also import `propBoolean`.